# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: MoleSignal API
  version: 1.0.0
  description: |
    MoleSignal 是可观测平台核心引擎。所有 endpoint 默认走 `Authorization: Bearer <jwt>` 鉴权；
    `/api/v1/healthz` / `/api/v1/auth/signin` / `/api/v1/auth/sso/*` / `/metrics`
    在白名单内免鉴权。

    错误统一返 `{ "error": "<message>" }` body；状态码：
    - 400 invalid argument
    - 401 unauthorized (token 校验失败)
    - 403 forbidden (IAM permission / license 不足)
    - 404 not found / 跨 org lookup
    - 409 conflict (unique 约束 / 状态冲突)
    - 501 profiling kind unavailable on this build
    - 413 quota exhausted
    - 429 rate limited

    精确流名 `_molesignal` 保留给服务自身遥测。所有公共 ingest 与 stream mutation
    对该名称返回 403；`_sys` system scope 中的 IAM 查询仍然可用。
servers:
  - url: https://molesignal.example.com
  - url: http://localhost:5080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/signin:
    post:
      summary: Email + password login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user_id: { type: string }
                  org_id: { type: string }
                  display_role: { type: string }
                  roles:
                    type: array
                    items:
                      type: object
                      required: [id, key, name, builtin]
                      properties:
                        id: { type: string }
                        key: { type: string }
                        name: { type: string }
                        builtin: { type: boolean }
        '401': { $ref: '#/components/schemas/Error' }
...
  /api/v1/auth/sso/ldap/login:
    post:
      summary: LDAP search and user simple-bind login (license-gated)
      description: |
        Resolves exactly one user DN with the configured escaped search filter, binds as that
        user over LDAPS or StartTLS, applies the configured identity-field and group-to-role
        mappings, provisions the local membership, and issues a local JWT.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_id, username, password]
              properties:
                provider_id: { type: string }
                username: { type: string, maxLength: 320 }
                password: { type: string, format: password, maxLength: 1024 }
      responses:
        '200': { description: Local JWT and organization context issued }
        '401': { description: Provider, user, or credentials invalid }
        '403': { description: sso feature not licensed }
        '503': { description: LDAP directory unavailable }
...
  /api/v1/_cloudflare:
    post:
      summary: Cloudflare Logpush push (NDJSON)
      parameters:
        - { in: header, name: X-Connector-Token, required: true, schema: { type: string } }
      responses:
        '204': { description: accepted }
  /api/v1/_heroku:
    post:
      summary: Heroku log drain push (text)
      parameters:
        - { in: header, name: X-Connector-Token, required: true, schema: { type: string } }
      responses:
        '204': { description: accepted }
...